+++ client/src/components/Header.jsx
@@ -1,8 +1,33 @@
 import Register from "./RegisterForm"
-// import { useCartContext } from "../context/CartContext";
-import { handlePayment } from "./Checkout";
-function Header() {
+import { useCartContext } from "../context/CartContext";
 
+function Header() {
+    const { cart } = useCartContext();
+    const transformedCart = cart.map(item => ({
+        product: item.product.default_price,
+        quantity: item.quantity // Assuming a default quantity of 1 for each product.
+    }));
+    async function handlePayment() {
+        const response = await fetch("http://localhost:3000/api/create-checkout-session", {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify(transformedCart),
+            credentials: 'include'
+        }
+        );
+        if (response.status === 401) {
+            alert("Your session has expired. Please log in again.");
+            window.location = "/"; // replace with your login page URL
+            return;
+        }
+        if (!response.ok) {
+            alert("Log in to procced!");
+            return
+        } const { url } = await response.json();
+        window.location = url;
+    }
     return (
         <div className="flex justify-between items-center bg-blue-500 p-4">
             <Register />
