+++ server/db/orders.json
@@ -393,5 +393,25 @@
     ],
     "paymentStatus": "paid",
     "paymentCompletedDate": "2023-09-11T19:36:07.000Z"
+  },
+  {
+    "customerId": "cus_Oc1x4HJD9gVkxj",
+    "customer": "Askarians@yahoo.com",
+    "amountTotal": 1304,
+    "currency": "sek",
+    "orderedItems": [
+      {
+        "name": "Juridik",
+        "price": 870,
+        "quantity": 1
+      },
+      {
+        "name": "Naturkunskap",
+        "price": 760,
+        "quantity": 1
+      }
+    ],
+    "paymentStatus": "paid",
+    "paymentCompletedDate": "2023-09-12T06:57:16.000Z"
   }
 ]
